use nomic-bitcoin enriched header type
