+++ core/primitives/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Judd <kepzorz@gmail.com>"]
 edition = "2018"
 
 [dependencies]
-bitcoin = "^0.19.1"
+nomic-bitcoin = { path = "../bitcoin"}
 serde = {version= "1.0.100", features=["derive"]}
